+++ litmus/index.js
@@ -106,7 +106,8 @@ import "./index.scss";
 //import Demo from "./bugs/SliderValue";
 //import Demo from "./features/context-menu/conext-menu-in-a-dropdown";
 // import Demo from "./features/calendar";
-import Demo from "./features/slider/SliderPreventDefault";
+//import Demo from "./features/slider/SliderPreventDefault";
+import Demo from "./features/validator/index";
 
 let store = (window.store = new Store());